[Driver] Print diags to stdout if --sema is passed
The `--sema` flag is mainly used for tests, which require any output
to be passed via stdout. We were previously appending `|& cat` to all
the run directives of the tests in question, but that’s not only jank
but also not portable, so this is a better solution to that.

This fixes #6