[Refactor] Remove explicit padding fields.
The original motivation for adding explicit padding fields to structs
is that field accesses in LLVM on structs are handled via indices; since
there is no way in LLVM to specify an alignment for a field, constructs
such as C++’s `alignas` necessiate the insertion of explicit padding
fields; the code that emits struct GEPs, in turn, needs to be aware
of this lest it accidentally emit an access to a padding fields instead.

There are two major problems with this approach:

  1. The frontend is absolutely the wrong place to put this; this is
     something that is only relevant in codegen; adding explicit padding
     fields to the AST means that code that uses RecordTypes now has to
     be made aware that they exist, and creating record types manually,
     e.g. when importing declarations from C++ or for a function’s
     static chain area in the backend, becomes unnecessarily tedious.

  2. There isn’t much value in doing struct accesses this way: LLVM
     itself is shifting away from typed GEPs in favour of a PTRADD
     instruction, and type information in GEPs has never really meant
     anything.

This removes padding fields from the frontend; we currently don’t
support overaligned fields, so once we do, the backend will have to
be updated accordingly.